moved search results as well as added transactions
